# Speedup CPU training by 10% using Memory Aligned Tensors